+++ modules/lambdas/variables.tf
@@ -1,12 +1,6 @@
 ## Variables obtained by Cumulus deployment
 ## Should exist in https://github.com/nasa/cumulus-template-deploy/blob/master/cumulus-tf/variables.tf
 ## REQUIRED
-variable "aws_profile" {
-  type        = string
-  description = "AWS profile used to deploy the terraform application."
-}
-
-
 variable "buckets" {
   type        = map(object({ name = string, type = string }))
   description = "S3 bucket locations for the various storage types being used."
@@ -34,13 +28,6 @@ variable "prefix" {
 variable "vpc_id" {
   type        = string
   description = "Virtual Private Cloud AWS ID"
-}
-
-
-## OPTIONAL - Default variable value is set in ../variables.tf to keep default values centralized.
-variable "region" {
-  type        = string
-  description = "AWS region to deploy configuration to."
 }