+++ modules/workflows/main.tf
@@ -1,21 +1,3 @@
-## Terraform Requirements
-terraform {
-  required_providers {
-    aws = {
-      source  = "hashicorp/aws"
-      version = ">= 3.5.0"
-    }
-  }
-}
-
-
-## AWS Provider Settings
-provider "aws" {
-  region  = var.region
-  profile = var.aws_profile
-}
-
-
 ## Local Variables
 locals {
   tags = merge(var.tags, { Deployment = var.prefix })
@@ -32,9 +14,7 @@ module "orca_recovery_workflow" {
   ## Cumulus Variables
   ## --------------------------
   ## REQUIRED
-  aws_profile     = var.aws_profile
   prefix          = var.prefix
-  region          = var.region
   system_bucket   = var.system_bucket
   tags            = local.tags
   workflow_config = var.workflow_config
@@ -58,9 +38,7 @@ module "orca_copy_to_glacier_workflow" {
   ## Cumulus Variables
   ## --------------------------
   ## REQUIRED
-  aws_profile     = var.aws_profile
   prefix          = var.prefix
-  region          = var.region
   system_bucket   = var.system_bucket
   tags            = local.tags
   workflow_config = var.workflow_config